+++ package.json
@@ -47,6 +47,7 @@
     "@nestjs/typeorm": "^10.0.0",
     "axios": "^1.5.0",
     "bcrypt": "^5.1.1",
+    "builder-pattern": "^2.2.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
     "cross-env": "^7.0.3",
