+++ src/api/comments/comments.module.ts
@@ -1,8 +1,8 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { Comments } from "src/entities/comments";
-import { CommentsController } from "./comments.controller";
-import { CommentsService } from "./comments.service";
+import { Comments } from "@entities/comments";
+import { CommentsController } from "@api/comments/comments.controller";
+import { CommentsService } from "@api/comments/comments.service";
 
 @Module({
   imports: [TypeOrmModule.forFeature([Comments])],
