+++ src/api/images/images.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { Images } from "src/entities/images";
-import { ImagesService } from "./images.service";
+import { Images } from "@entities/images";
+import { ImagesService } from "@api/images/images.service";
 
 @Module({
   imports: [TypeOrmModule.forFeature([Images])],
