+++ src/api/images/images.service.ts
@@ -1,9 +1,6 @@
-import { GetObjectCommand, S3Client } from "@aws-sdk/client-s3";
-import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
-import setting from "src/config/setting";
-import { Images } from "src/entities/images";
+import { Images } from "@entities/images";
 import { Repository } from "typeorm";
 
 @Injectable()
