+++ src/api/posts/posts.http
@@ -1,7 +1,7 @@
 ### 글 쓰기
 POST http://localhost:4000/api/posts
 Content-Type: application/json
-
+ 
 {
   "board_id": 1,
   "user_id": 1,
@@ -10,7 +10,7 @@ Content-Type: application/json
 }
 
 ### 글 조회
-GET http://localhost:4000/api/posts/2
+GET http://localhost:4000/api/posts/8
 
 ### 글 수정
 PUT http://localhost:4000/api/posts/7
