+++ src/api/posts/posts.module.ts
@@ -1,10 +1,10 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { Posts } from "src/entities/posts";
-import { PostsController } from "./posts.controller";
-import { PostsService } from "./posts.service";
-import { Images } from "src/entities/images";
-import { ImagesService } from "../images/images.service";
+import { Posts } from "@entities/posts";
+import { PostsController } from "@api/posts/posts.controller";
+import { PostsService } from "@api/posts/posts.service";
+import { Images } from "@entities/images";
+import { ImagesService } from "@api/images/images.service";
 
 @Module({
   imports: [TypeOrmModule.forFeature([Posts, Images])],
