+++ src/api/posts/posts.service.spec.ts
@@ -1,6 +1,6 @@
 // user.service.spec.ts
 import { Test, TestingModule } from "@nestjs/testing";
-import { PostsService } from "./posts.service";
+import { PostsService } from "@api/posts/posts.service";
 
 describe("PostsService", () => {
   let service: PostsService;
