+++ src/dto/comments/create-comment.dto.ts
@@ -1,4 +1,4 @@
 import { PickType } from "@nestjs/swagger";
-import { Comments } from "src/entities/comments";
+import { Comments } from "@entities/comments";
 
 export class CreateCommentDto extends PickType(Comments, ["parent_id", "post_id", "user_id", "depth", "content"]) {}
