+++ src/dto/comments/update-comment.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from "@nestjs/swagger";
-import { CreateCommentDto } from "./create-comment.dto";
+import { CreateCommentDto } from "@dto/comments/create-comment.dto";
 
 export class UpdateCommentDto extends PartialType(CreateCommentDto) {}
