+++ src/dto/posts/post.response.ts
@@ -1,8 +1,8 @@
 import { PickType } from "@nestjs/swagger";
 import { Exclude, plainToClass } from "class-transformer";
 import { isArray } from "class-validator";
-import { Posts } from "src/entities/posts";
-import { Users } from "src/entities/users";
+import { Posts } from "@entities/posts";
+import { Users } from "@entities/users";
 
 export class PostResponse extends PickType(Posts, ["id", "board_id", "user_id", "title", "content", "writer", "created_at", "updated_at"]) {
   @Exclude()
