+++ src/dto/posts/update-post.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from "@nestjs/swagger";
-import { CreatePostDto } from "./create-post.dto";
+import { CreatePostDto } from "@dto/posts/create-post.dto";
 
 export class UpdatePostDto extends PartialType(CreatePostDto) {}
