+++ src/entities/images.ts
@@ -10,11 +10,9 @@ import {
   CreateDateColumn,
   AfterLoad,
 } from "typeorm";
-import { Posts } from "./posts";
-import { DateProperty, NumProperty, ObjectProperty, StringProperty } from "src/common/decorators/common/property.decorator";
-import setting from "src/config/setting";
-import { GetObjectCommand, S3Client } from "@aws-sdk/client-s3";
-import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
+import { Posts } from "@entities/posts";
+import { DateProperty, NumProperty, ObjectProperty, StringProperty } from "@common/decorators/common/property.decorator";
+import { getS3SignedUrl } from "@utils/custom-s3-client";
 
 @Entity()
 export class Images extends BaseEntity {
@@ -56,17 +54,6 @@ export class Images extends BaseEntity {
 
   @AfterLoad()
   async toDisplay() {
-    const s3 = new S3Client({
-      region: setting.AWS.REGION,
-      credentials: {
-        accessKeyId: setting.AWS.ACCESS_KEY_ID,
-        secretAccessKey: setting.AWS.SECRET_ACCESS_KEY,
-      },
-    });
-
-    const getObjectParams: any = { Bucket: this.bucketname, Key: this.filename };
-    const command: any = new GetObjectCommand(getObjectParams);
-
-    this.signedUrl = await getSignedUrl(s3, command, { expiresIn: 60 * 5 }); // 5min
+    this.signedUrl = await getS3SignedUrl(this.bucketname, this.filename);
   }
 }
